common/hooks: Fix hooks when build_style=meta

+++ common/hooks/post-install/02-remove-python-bytecode-files.sh
@@ -1,5 +1,7 @@
 # This hook removes python bytecode files (.py[co]).
 
 hook() {
-	find ${PKGDESTDIR} -type f -name '*.py[co]' -delete
+    if [ -d "${PKGDESTDIR}" ]; then
+        find ${PKGDESTDIR} -type f -name '*.py[co]' -delete
+    fi
 }
